style: Corner buttons outside of header
To keep them in upper right corner even if the app is vertically centered
