+++ src/components/HeroHeader/index.jsx
@@ -2,32 +2,6 @@ import React from 'react'
 import { useClient } from 'cozy-client'
 import get from 'lodash/get'
 import useInstanceSettings from 'hooks/useInstanceSettings'
-import flag from 'cozy-flags'
-
-import LogoutButton from './LogoutButton'
-import SettingsButton from './SettingsButton'
-import HelpButton from './HelpButton'
-
-const cornerButtons = [
-  {
-    flagName: 'help-is-displayed',
-    isDisplayedByDefault: true,
-    Button: HelpButton
-  },
-  {
-    flagName: 'settings-is-displayed',
-    isDisplayedByDefault: false,
-    Button: SettingsButton
-  },
-  {
-    flagName: 'logout-is-displayed',
-    isDisplayedByDefault: true,
-    Button: LogoutButton
-  }
-]
-
-const flagWithFallbackValue = (flagName, fallback) =>
-  flag(flagName) === null ? fallback : flag(flagName)
 
 export const HeroHeader = () => {
   const client = useClient()
@@ -39,16 +13,6 @@ export const HeroHeader = () => {
 
   return (
     <header className="hero-header">
-      <div className="corner">
-        {cornerButtons.map(({ flagName, isDisplayedByDefault, Button }) =>
-          flagWithFallbackValue(
-            `home.corner.${flagName}`,
-            isDisplayedByDefault
-          ) ? (
-            <Button key={flagName} />
-          ) : null
-        )}
-      </div>
       <div>
         <img className="hero-avatar" src={`${rootURL}/public/avatar`} />
       </div>
