+++ src/styles/app.styl
@@ -27,6 +27,7 @@ body
 
 .App
     display flex
+    flex-direction column
     min-width 100%
     width 100%
     background-color var(--defaultBackgroundColor)
@@ -36,6 +37,18 @@ body
     mix-blend-mode multiply
     align-items center
     overflow-y auto
+    .corner
+        position relative
+        align-self flex-end
+        margin rem(12) 2rem rem(-44) 0
+        a, button
+            color: var(--white)
+            +small-screen()
+                height 3rem
+                width 3rem
+                margin 0
+        +small-screen()
+            margin 0 0 -3rem
 
 .App::before
     content ''
