+++ src/styles/hero-header.styl
@@ -15,13 +15,6 @@ whiteOnWhiteShadow = drop-shadow(0 1px 1px rgba(50, 54, 63, .32)) drop-shadow(0 
     +tiny-screen()
         padding 3rem 4rem 2rem
 
-    .corner
-        position absolute
-        top rem(8)
-        right rem(8)
-        a, button
-            color: var(--regularButtonConstrastColor)
-
     .hero-avatar
         width 8rem
         height 8rem
